pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-code-repair"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pytest>=7.4", "google-genai>=1.0", "python-dotenv>=1.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"